Removed unnacessary files
